pseudo: Implemented means to force chroot usage
This is for executables with exceptions and special needs.

Hmm. I also removed one unused variable now that I should have removed
in a prior commit.

The idea here is you may want to be able to run specific things from
the chroot because neither host nor sysroot provides it properly. You
are asking about the difference in architecture for scripts and target
binaries? Just wait for a subsequent commit

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>