seudo: Now handles chroot scripts and executables
We rewrite paths and arguments as needed to enable chroot execution.
The problem is that argv0 needs to be rewritten properly for chroot
environments so that the file is found by any interpreter, even if
that interpreter is not searching PATH ... which most don't.

This patch also adds in a buffer size correction and check that
Catalin provided.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>
Signed-off-by: Costea Ciprian <ciprianmarian.costea@nxp.com>