pseudo did not treat softlinks correctly in chroot
When decoding a softlink inside a chroot environment, the
chroot exceptions were not taken into account. When, e.g.,
/dev/null is an exception to chroot handling to get the host
side null device, then a softlink pointing to such a target
also needs to respect the configured exception.
Turns out that we need to decode softlinks for an exec path
differently than for a file path because exception handling
is slightly different.
This patch fixes running systemctl in image.bbclass for a
complex rootfs and likely some other things.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>