pseudo: Fix fxstatat*() calls for non fs file handles
This is a really ugly one because it shows a deficiency in how
pseudo currently handles paths. When you fstatat, e.g., stdout,
there may not be a real path going with it. This means that
pseudo_root_path() via base_path() converts the original path "" to
0. There are two problems with that. The real functions still want to
see the "" instead of the 0 or they return EFAULT which confuses
callers when they just expect to check on stdout redirection.
The second problem is that there is no differentiation between an
error returned by base_path and a 0 path. We also can't just go from
0 to "" in that case because that messes up our path databases as the
code is not currently designed to deal with that. To really properly
differentiate between pseudo path vs real path vs 0 vs error needs
much more thought and smart ideas than I currently have.

So for now we fix it with a hack. First, we second guess base_path to
determine if the 0 really is an error or not to not print spurious
diagnostics when a 0 return may well be legitimate.
Second, we mutilate the two fstatat*() variants to convert back from
0 to "" so that the function works. Technically this changes the
semantics of the wrappers but it isn't all bad as pseudo would have
checked 0 paths earlier even.

All in all, I am not really happy with this patch, but it will do the
job and while ugly at least it isn't overly obscure or clever.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>